Split volume-related cli methods into module
